+++ scripts/custom_tree.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+find . -path '*/.git*' -prune -o -print |
+	sed -e 's;[^/]*/;|___;g;s;___|; |;g'