+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### master
+
+### v0.2.0, Sep 03, 2014
 - automatic sidebar opening for another sidebar type
 - sidebars can't have nested sidebars
 - show "custom tree" dir listing when `tree` command is not installed